refactor: extract sync blob helpers
